+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## Next
 
 * Fix: site-name of docs does not follow renaming
+* Extract flatten_latex to its own package [latex-include](https://pypi.org/project/latex-include/)
 
 
 ## 0.4.0